app/requirements.txt
backoff==2.2.1
blinker==1.7.0
click==7.1.2
Deprecated==1.2.14
Flask==1.1.4
Flask-OpenTracing==1.1.0
googleapis-common-protos==1.59.1
grpcio==1.60.1
importlib-metadata==6.8.0
install==1.3.5
itsdangerous==1.1.0
jaeger-client==4.8.0
Jinja2==2.11.3
MarkupSafe==2.0.1
opentelemetry-api==1.22.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opentracing==2.4.0
packaging==23.2
protobuf==4.25.2
six==1.16.0
threadloop==1.0.2
thrift==0.16.0
tornado==6.4
typing_extensions==4.9.0
Werkzeug==1.0.1
wrapt==1.16.0
zipp==3.17.0
